refactor: extract `uploadConflictHandler` from `conflicts.ts`
