+++ __tests__/utils/conflicts-handler.spec.ts
@@ -4,7 +4,7 @@
  */
 
 import { describe, expect, it, vi, beforeEach } from 'vitest'
-import { uploadConflictHandler } from '../../lib/utils/conflicts.ts'
+import { uploadConflictHandler } from '../../lib/ui/utils/uploadConflictHandler.ts'
 import { InvalidFilenameError, InvalidFilenameErrorReason, File as NcFile } from '@nextcloud/files'
 
 const validateFilename = vi.hoisted(() => vi.fn(() => true))
