+++ lib/ui/utils/uploadConflictHandler.ts
@@ -1,45 +1,20 @@
-/**
+/*!
  * SPDX-FileCopyrightText: 2024 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
 import type { Node } from '@nextcloud/files'
-import type { IDirectory } from '../core/utils/fileTree.ts'
+import type { IDirectory } from '../../core/utils/fileTree.ts'
 
 import { showInfo, showWarning } from '@nextcloud/dialogs'
 import { getUniqueName, InvalidFilenameError, validateFilename } from '@nextcloud/files'
 import { basename } from '@nextcloud/paths'
 
-import { openConflictPicker } from '../index.ts'
-import { showInvalidFilenameDialog } from '../ui/utils/dialog.ts'
-import { t } from '../shared/utils/l10n.ts'
-import logger from '../shared/utils/logger.ts'
-
-/**
- * Check if there is a conflict between two sets of files
- * @param {Array<File|FileSystemEntry|Node>} files the incoming files
- * @param {Node[]} content all the existing files in the directory
- * @return {boolean} true if there is a conflict
- */
-export function hasConflict(files: (File|FileSystemEntry|Node)[], content: Node[]): boolean {
-	return getConflicts(files, content).length > 0
-}
-
-/**
- * Get the conflicts between two sets of files
- * @param {Array<File|FileSystemEntry|Node>} files the incoming files
- * @param {Node[]} content all the existing files in the directory
- * @return {boolean} true if there is a conflict
- */
-export function getConflicts<T extends File|FileSystemEntry|Node>(files: T[], content: Node[]): T[] {
-	const contentNames = content.map((node: Node) => node.basename)
-	const conflicts = files.filter((node: File|FileSystemEntry|Node) => {
-		const name = 'basename' in node ? node.basename : node.name
-		return contentNames.indexOf(name) !== -1
-	})
-
-	return conflicts
-}
+import { getConflicts } from '../../core/utils/conflicts.ts'
+import { openConflictPicker } from '../../index.ts'
+import { showInvalidFilenameDialog } from './dialog.ts'
+import { t } from '../../shared/utils/l10n.ts'
+import logger from '../../shared/utils/logger.ts'
 
 /**
  * Helper function to create a conflict resolution callback for the `Uploader.batchUpload` method.
